battery level bug fixed
Before the battery level is sent as Average even if the battery is full (low battery level is working fine), here LVD 3v3 causing an issue.
Now code flow for checking the battery level is changed. 1st it will check for battery full status and if a value is greater than the full battery threshold it will send the full indication, else it will check for the 3V0 LVD flag, if the 3v0 flag is high it will send a low battery else it will send average value.